lwdOuter defaults to 0.5 instead of 0.
With the previous default of lwdOuter at 0 no line was shown, although the
definition of the outerCI was 2, and it did not make much sense to define the
width of the outer interval but not have it plotted. With this new default a
line half the width of the inner will be shown.